+++ src/main/kotlin/no/nav/familie/ef/sak/opplysninger/personopplysninger/endringer/EndringerIPersonopplysningerDto.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ef.sak.opplysninger.personopplysninger.dto
+package no.nav.familie.ef.sak.opplysninger.personopplysninger.endringer
 
 import java.time.LocalDateTime
 
@@ -55,23 +55,3 @@ data class Endringer(
 }
 
 data class Endring(val harEndringer: Boolean = false)
-
-fun finnEndringer(tidligere: PersonopplysningerDto, nye: PersonopplysningerDto) =
-    Endringer(
-        folkeregisterpersonstatus = utledEndringer(tidligere.folkeregisterpersonstatus, nye.folkeregisterpersonstatus),
-        fødselsdato = utledEndringer(tidligere.fødselsdato, nye.fødselsdato),
-        dødsdato = utledEndringer(tidligere.dødsdato, nye.dødsdato),
-        statsborgerskap = utledEndringer(tidligere.statsborgerskap, nye.statsborgerskap),
-        sivilstand = utledEndringer(tidligere.sivilstand, nye.sivilstand),
-        adresse = utledEndringer(tidligere.adresse, nye.adresse),
-        fullmakt = utledEndringer(tidligere.fullmakt, nye.fullmakt),
-        barn = utledEndringer(tidligere.barn, nye.barn), // TODO bedre diff på barn
-        // andreForeldre = // TODO legge til andre forelder og adresse på disse?
-        innflyttingTilNorge = utledEndringer(tidligere.innflyttingTilNorge, nye.innflyttingTilNorge),
-        utflyttingFraNorge = utledEndringer(tidligere.utflyttingFraNorge, nye.utflyttingFraNorge),
-        oppholdstillatelse = utledEndringer(tidligere.oppholdstillatelse, nye.oppholdstillatelse),
-        vergemål = utledEndringer(tidligere.vergemål, nye.vergemål),
-    )
-
-private fun <T> utledEndringer(tidligere: T, nye: T) =
-    Endring(tidligere != nye)
